graphql-api/requirements.txt
boto3==1.22.9
fastapi==0.76.0
pydantic==1.9.0
pymongo==4.1.1
strawberry-graphql[fastapi]==0.114.0
uvicorn==0.17.6
